Add files via upload
Added changes which makes magicsoundsystem.dll into magicsystem.dll, which adds functionality for Screen Refresh, Event Handling and Timing.

Please note (as I noticed with "Secret Game Project #1" - for which these changes were included - the timing stuff does not work like this, put into an external DLL - no idea why - which is why in Secret Project #1 I reimplemented the timing code to only use MSS_GetTicks() and not AddTimer/RemoveTimer. The reason for the two code paths for MSS_GetTicks is that I now also added an AGA Version, and the SDL AGA does not include SDL_GetTicks().

The OS4 version sort of caused trouble when trying to dlopen it (I linked it static to Secret Project #1, which is no problem there, as SDL2 has a nice licence). For 68k the dynamic linkage worked (I use three versions of the DLL, one for graphics board systems, one for AGA and one for AGA without FPU -possibly in the future also one for WarpOS).